synch to master

+++ artio-codecs/src/main/java/uk/co/real_logic/artio/otf/OtfMessageAcceptor.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2015-2017 Real Logic Ltd.
+ * Copyright 2015-2020 Real Logic Limited.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ * https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -56,5 +56,5 @@ public interface OtfMessageAcceptor
      */
     MessageControl onGroupEnd(int tag, int numInGroup, int index);
 
-    boolean onError(ValidationError error, int messageType, int tagNumber, AsciiFieldFlyweight value);
+    boolean onError(ValidationError error, long messageType, int tagNumber, AsciiFieldFlyweight value);
 }
